file-service/requirements.txt
databases[postgresql]==0.2.6
fastapi==0.65.0
SQLAlchemy==1.3.13
uvicorn==0.11.2
httpx==0.11.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
asyncpg==0.20.1
python-multipart
pydantic[email]
requests
aiofiles
pyctuator
psutil